+++ src/edu/upenn/cit594/util/RealEstateAverageableAttributeFetcher.java
@@ -2,6 +2,6 @@ package edu.upenn.cit594.util;
 
 public abstract class RealEstateAverageableAttributeFetcher {
 
-    public abstract int fetchAttribute(Property p);
+    public abstract double fetchAttribute(Property p);
 
 }
